+++ utils.py
@@ -1,7 +1,10 @@
+import asyncio
 import gzip
+import typing
 
+import aiofiles
 import virtool_core.history.db
-
+import virtool_core.otus.utils
 
 ISOLATE_KEYS = [
     "id",
@@ -29,6 +32,65 @@ SEQUENCE_KEYS = [
     "host",
     "sequence"
 ]
+
+
+async def get_patched_otus(db, data_path, manifest):
+    coros = list()
+
+    for patch_id, patch_version in manifest.items():
+        coros.append(virtool_core.history.db.patch_to_version(
+            db,
+            data_path,
+            patch_id,
+            patch_version
+        ))
+
+    return [j[1] for j in await asyncio.tasks.gather(*coros)]
+
+
+async def write_sequences_to_file(path, sequences):
+    """
+    Write a FASTA file based on a given `Iterable` containing sequence documents.
+    Headers will contain the `_id` field of the document and the sequence text is from the `sequence` field.
+    :param path: the path to write the file to
+    :param sequences: the sequences to write to file
+    """
+    async with aiofiles.open(path, "w") as f:
+        for sequence in sequences:
+            sequence_id = sequence["_id"]
+            sequence_data = sequence["sequence"]
+
+            line = f">{sequence_id}\n{sequence_data}\n"
+            await f.write(line)
+
+
+def get_sequences_from_patched_otus(
+        otus: typing.Iterable[dict],
+        data_type: str, sequence_otu_map: dict
+) -> typing.Generator[dict, None, None]:
+    """
+    Return sequence documents based on an `Iterable` of joined OTU documents. Writes a map of sequence IDs to OTU IDs
+    into the passed `sequence_otu_map`.
+    If `data_type` is `barcode`, all sequences are returned. Otherwise, only sequences of default isolates are returned.
+    :param otus: an Iterable of joined OTU documents
+    :param data_type: the data type of the parent reference for the OTUs
+    :param sequence_otu_map: a dict to populate with sequence-OTU map information
+    :return: a generator that yields sequence documents
+    """
+    for otu in otus:
+        otu_id = otu["_id"]
+
+        for isolate in otu["isolates"]:
+            for sequence in isolate["sequences"]:
+                sequence_id = sequence["_id"]
+                sequence_otu_map[sequence_id] = otu_id
+
+        if data_type == "barcode":
+            for sequence in virtool_core.otus.utils.extract_sequences(otu):
+                yield sequence
+        else:
+            for sequence in virtool_core.otus.utils.extract_default_sequences(otu):
+                yield sequence
 
 
 async def export(db, data_path, ref_id):
